outsourced PoweredByUtil
